删除收获地址

+++ app/Http/Controllers/UserAddressesController.php
@@ -47,6 +47,8 @@ class UserAddressesController extends Controller
      */
     public function edit(UserAddress $userAddress)
     {
+        $this->authorize('own', $userAddress);
+
         return view('user_addresses.create_and_edit', ['address' => $userAddress]);
     }
 
@@ -55,6 +57,8 @@ class UserAddressesController extends Controller
      */
     public function update(UserAddressRequest $request, UserAddress $userAddress)
     {
+        $this->authorize('own', $userAddress);
+
         $userAddress->update($request->only([
             'province',
             'city',
@@ -70,9 +74,10 @@ class UserAddressesController extends Controller
     /**
      * 删除收获地址
      */
-    public function distroy(UserAddress $userAddress)
+    public function destroy(UserAddress $userAddress)
     {
+        $this->authorize('own', $userAddress);
         $userAddress->delete();
-        return redirect()->route('user_addresses.index');
+        return [];
     }
 }
